begin creation of RT

+++ processManager.c
@@ -5,6 +5,7 @@
 
 #include <unistd.h>
 #include <sys/wait.h>
+#include <fcntl.h>
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -18,22 +19,23 @@
 
 ProcInfo * getProcInfo(int fd[]);
 
-
-
-
 void AlrmHandler(int sinal);
+int afterRTProc(ProcInfo *);
 
 int flag = 0; //flag que diz se o processo deve escalonar ou olhar a pipe
 
+
 int main (int argc, char *argv[]) { 
 	
-	int pid, i,fd[2], sizeOfProcInfo;
+	int pid, i,fd[2], sizeOfProcInfo, realTime[60];
 	char serializedProcInfo[100];
-
+	
 	if(pipe(fd) < 0){
 			printf("Erro ao criar pipe");
 			exit(1);		
 	}
+	fcntl(fd[0], F_SETFL, O_NONBLOCK); //Set read fd as Non-Blocking
+	
 	
 	for(i=0; i<3 ;i++){
 		if((pid = fork()) < 0){
@@ -66,42 +68,43 @@ int main (int argc, char *argv[]) {
 			signal(SIGALRM, AlrmHandler); 
 			while(1){
 				if(flag==0) /* Pode se inserir um novo processo na lista de processos*/ {
-					
-					procInfo = getProcInfo(fd);									//precisamos fazer isso ser non-blocking (trocar pra FIFO talvez?)
-					printf("Nome do Processo: %s\n", procInfo->nomeProc);
-					printf("Tipo do Processo: %s\n", procInfo->tipoProc);
-					
-					char *nome = (char*) malloc (sizeof(char)*(3+strlen(procInfo->nomeProc))); 
-					nome[0] = '.';
-					nome[1] = '/';
-					nome[2] = '\0';
-					
-					strcat(nome,procInfo->nomeProc);
-					if ((pidProc = fork()) == 0) /* Cria o novo processo */ {
-						char *arg[] = {nome,NULL};
-						execv(arg[0], arg);
-					}
-					else{
-						kill(pidProc, SIGSTOP); //interrompe o recém nascido
-						if(procInfo->I==NULL) /* Prioridade ou Round Robin */ {
-							int prioridade;
-							if(procInfo->PR == NULL) /* Round Robin */ {prioridade = 8;}
-							else /* Prioridade */{prioridade = atoi(procInfo->PR);}
-							listaProcs = insereElemento(listaProcs, pid, prioridade);  //este insereElemento vai inserir na posição correta (?)
-						}
-						else /* Real Time*/ {
+					if((procInfo = getProcInfo(fd)) != NULL) /* Há um processo novo*/ {
+						printf("Nome do Processo: %s\n", procInfo->nomeProc);
+						printf("Tipo do Processo: %s\n", procInfo->tipoProc);
 						
+						char *nome = (char*) malloc (sizeof(char)*(3+strlen(procInfo->nomeProc))); 
+						nome[0] = '.';
+						nome[1] = '/';
+						nome[2] = '\0';
 						
+						strcat(nome,procInfo->nomeProc);
+						if ((pidProc = fork()) == 0) /* Cria o novo processo */ {
+							char *arg[] = {nome,NULL};
+							execv(arg[0], arg);
+						}
+						else{
+							kill(pidProc, SIGSTOP); //interrompe o recém nascido
+							if(procInfo->I==NULL) /* Prioridade ou Round Robin */ {
+								int prioridade;
+								if(procInfo->PR == NULL) /* Round Robin */ {prioridade = 8;}
+								else /* Prioridade */{prioridade = atoi(procInfo->PR);}
+								listaProcs = insereElemento(listaProcs, pid, prioridade);  //este insereElemento vai inserir na posição correta (?)
+							}
+							else /* Real Time*/ {
+							
+								if(afterRTProc(procInfo)) /* Process starts after another Real Time process */ {
+								
+								}
+								else /* Process starts at arbitrary time */{
+								
+								}
+							}
 						}
 					}
-					
 				}
-				else /* Um processo acabou de ser interrompido (é para isso que serve esse bloco né ?? || honestamente não me lembro mais ) */ {
+				else /* Um processo acabou de ser interrompido */ {
 					
-					/*
-					OBS:ISTO DEVE SER ENVOLTO EM UM MUTEX, PARA QUE NAO SEJA INTERROMPIDO CASO HAJA UM NOVO SIGALRM
-					O SIGALRM TMB DEVE SER ENVOLTO NESTE MUTEX, PARA QUE AGUADE O FIM DA EXECUÇÃO DESTE BLOCO
-					
+					/*				
 					
 					
 					
@@ -113,7 +116,12 @@ int main (int argc, char *argv[]) {
 		}
 		
 		else if(i == 2) /* processo 3 */{
+					/*
+					Este processo cuida do escalonamento
 					
+						-Deve enviar alarmes ao processo 2 quanto um processo for interrompido
+					
+					*/
 		} 
 	}
 	else{
@@ -129,18 +137,21 @@ int main (int argc, char *argv[]) {
 ProcInfo * getProcInfo(int fd[]){
 	int sizeOfProcInfo;
 	char serializedProcInfo[100];
-	ProcInfo *procInfo = (ProcInfo*)malloc(sizeof(ProcInfo));
+	ProcInfo *procInfo ;
 
 	printf("Estou lendo do pipe\n");
-	read(fd[0],&sizeOfProcInfo,sizeof(int));
-	read(fd[0],serializedProcInfo,sizeOfProcInfo);
-	
-	
-	//read(fd[0],&sizeOfProcInfo,sizeof(int));
-	//read(fd[0],serializedProcInfo,sizeOfProcInfo);
+	if((read(fd[0],&sizeOfProcInfo,sizeof(int))) > 0 ) /* Há informação na pipe*/ {
+		read(fd[0],serializedProcInfo,sizeOfProcInfo);		
+		
+		//read(fd[0],&sizeOfProcInfo,sizeof(int));
+		//read(fd[0],serializedProcInfo,sizeOfProcInfo);
 
-	procInfo = deserializeProcInfo(serializedProcInfo);
-	return procInfo;
+		procInfo = deserializeProcInfo(serializedProcInfo);
+		return procInfo;
+	}
+	else{
+		return NULL;
+	}
 }
 
 
@@ -148,4 +159,13 @@ void AlrmHandler(int sinal)
 {
 	flag = 1;
 }
+
+int afterRTProc(ProcInfo * procInfo){
+	char *c;
+	for(c=procInfo->I ; *c ; c++){
+		if(*c < 48 || *c > 57) return 0;
+	}
+	return 1;
+}
+
